#54 New trait helper for tests, update to user files, companyindustry test, and routes
